scripts/requirements.txt
absl-py==0.13.0
adal==1.2.7
alembic==1.4.1
altair==4.2.0
aniso8601==9.0.1
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bitcointrade==0.7.0
bkcharts==0.2
black==19.10b0
bleach==4.1.0
blinker==1.4
boto==2.49.0
Bottleneck==1.3.2
Brlapi==0.7.0
Brotli==1.0.9
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.2.2
cassandra-driver==3.25.0
catboost==0.26.1
certifi==2021.10.8
cffi==1.14.6
chardet==3.0.4
cheap-repr==0.4.6
chrome-gnome-shell==0.0.0
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.3
colored==1.4.3
command-not-found==0.3
commonmark==0.9.1
compizconfig-python==0.9.14.1
config-checker==0.0.1
contextlib2==0.6.0.post1
cql==1.4.0
cqlsh==4.1.1
cryptography==2.8
ctypes-callable==1.0.4
cupshelpers==1.0
cycler==0.10.0
cytoolz==0.11.0
dash==1.17.0
dash-bootstrap-components==0.9.1
dash-core-components==1.13.0
dash-daq==0.5.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
dask==2022.1.1
databricks-cli==0.15.0
db-ops==0.1.1
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.7.1
dill==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.8
django-crispy-forms==1.13.0
django-htmlmin==0.11.0
dnspython==1.16.0
docker==5.0.0
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.2
Faker==16.7.0
fasteners==0.14.1
feather-format==0.4.1
Flask==1.1.2
Flask-Compress==1.10.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==2.0
fonttools==4.25.0
freeze-requirements==0.5.3
fsspec==2022.1.0
future==0.18.2
gast==0.5.3
geomet==0.2.1.post1
gitdb==4.0.7
GitPython==3.1.18
google==3.0.0
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg===1.13.1-unknown
graphviz==0.17
greenlet==1.1.2
grpcio==1.44.0
Guake==3.6.3
h5py==3.6.0
hijra==1.0
html5lib==1.1
htmlmin==0.1.12
httplib2==0.14.0
icecream==2.1.3
idna==2.8
imbalanced-learn==0.10.1
importlib-resources==5.4.0
influxdb==5.3.1
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.1
joblib==1.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kazam==1.4.5
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
linecache2==1.0.0
locket==0.2.1
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.5
Markdown==3.1.1
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
msgpack==1.0.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
notebook==6.4.8
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
partd==1.2.0
pathspec==0.7.0
pbr==5.4.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.4.1
plotly==4.12.0
powerline-status==2.7
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.20.0
psutil==5.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.11.2
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.21.1
python-editor==1.0.4
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
ranger-fm==1.9.3
regex==2022.7.25
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retrying==1.3.3
rich==12.6.0
rsa==4.8
scikit-learn==1.2.1
scipy==1.8.1
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
selenium==4.0.0a1
selenium-stealth==1.0.6
Send2Trash==1.8.0
sh==1.14.3
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smmap==4.0.0
snoop==0.4.3
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlparse==0.4.2
ssh-import-id==5.10
stack-data==0.1.4
sweetviz==2.1.4
systemd-python==234
tabulate==0.8.10
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
testresources==2.0.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
thrift==0.16.0
toml==0.10.2
tomli==2.0.0
toolz==0.11.2
tornado==6.1
tqdm==4.64.1
traceback2==1.4.0
traitlets==5.1.1
typed-ast==1.5.4
typing-extensions==4.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
usb-creator==0.3.7
varname==0.10.0
wadllib==1.3.3
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.1
widgetsnbextension==3.5.2
wrapt==1.14.0
xkit==0.0.0
youtube-dl==2021.12.17
zipp==3.7.0
